Move code utilities to the top level
